src/requirements.txt
binance-futures-connector
pandas
numpy